feat : refactored UI into multiple utility and page files, fixed a bug leading to OOM on big series of images, fixed a bug leading to uncorrectly stretched images
