+++ ui/__init__.py
@@ -1,0 +1,11 @@
+"""UI helper package for LightDiffusion-Next.
+
+This package contains small modules that encapsulate settings, history
+and rendering helpers so the main `streamlit_app.py` can remain focused
+on page composition and control flow.
+"""
+__all__ = [
+    "settings",
+    "history",
+    "helpers",
+]
